+++ src/app/store.ts
@@ -10,7 +10,7 @@ import { authApi } from '../features/auth';
 import { noticeApi } from '../features/notice';
 import { signupApi } from '../features/signup';
 import { classApi } from '../features/shop/shop-management/api/class-api';
-import { storyApi } from '../features/shop/shop-management/api/story-api';
+import { businessHoursApi } from '../features/shop/shop-management/api/business-hours-api';
 import { mapApi, oneDayClassApi } from '../features/map';
 import { chatApi } from '../features/chat/api/chat-api';
 import { myPageApi } from '../features/my-page/api/my-page-api';
@@ -31,7 +31,7 @@ export const store = configureStore({
     [noticeApi.reducerPath]: noticeApi.reducer,
     [signupApi.reducerPath]: signupApi.reducer,
     [classApi.reducerPath]: classApi.reducer,
-    [storyApi.reducerPath]: storyApi.reducer,
+    [businessHoursApi.reducerPath]: businessHoursApi.reducer,
     [mapApi.reducerPath]: mapApi.reducer,
     [oneDayClassApi.reducerPath]: oneDayClassApi.reducer,
     [chatApi.reducerPath]: chatApi.reducer,
@@ -48,7 +48,7 @@ export const store = configureStore({
       .concat(noticeApi.middleware)
       .concat(signupApi.middleware)
       .concat(classApi.middleware)
-      .concat(storyApi.middleware)
+      .concat(businessHoursApi.middleware)
       .concat(mapApi.middleware)
       .concat(oneDayClassApi.middleware)
       .concat(chatApi.middleware)
