+++ src/features/shop/shop-management/constants/business-hours-constants.ts
@@ -1,3 +1,3 @@
-export const STORY_ENDPOINTS = {
+export const BUSINESS_HOURS_ENDPOINTS = {
   BUSINESS_HOURS: '/api/v1/shops/me/business-hours',
 } as const;
