+++ src/features/shop/shop-management/index.ts
@@ -1,4 +1,4 @@
 export { ShopManagement } from '../shop-management/pages/shop-management';
 
 export { classApi } from '../shop-management/api/class-api';
-export { storyApi } from '../shop-management/api/story-api';
+export { storyApi } from './api/business-hours-api';
